behebe fehler wo int statt long stand

+++ src/main/kotlin/datamodel/device.kt
@@ -3,5 +3,5 @@ package datamodel
 data class device(
     var deviceName :String,
     var deviceId:String,
-    var functionIds:ArrayList<Int>
+    var functionIds:ArrayList<Long>
 )
